finalizing migration
