+++ packages/client/.env.local
@@ -1,0 +1,1 @@
+PUBLIC_SERVER_URL=http://localhost:3001
