+++ packages/client/index.html
@@ -106,7 +106,6 @@
     </svg>
   </div>
   <div id="root" class="app-hidden"></div>
-  <script type="module" src="/src/main.tsx"></script>
 </body>
 
 </html>
