+++ packages/client/tsconfig.node.json
@@ -6,5 +6,5 @@
 		"moduleResolution": "bundler",
 		"allowSyntheticDefaultImports": true
 	},
-	"include": ["./vite.config.ts", "./vitest.setup.ts"]
+	"include": ["./rsbuild.config.ts", "./vitest.setup.ts"]
 }
