+++ packages/client/vitest.config.ts
@@ -2,10 +2,8 @@
 
 import { defineConfig, mergeConfig } from "vitest/config";
 
-import viteConfig from "./vite.config";
-
 export default mergeConfig(
-	viteConfig,
+	{},
 	defineConfig({
 		test: {
 			globals: true,
